backend/pyproject.toml
[project]
name = "consulting-ai-backend"
version = "0.1.0"
description = "FastAPI backend for Consulting AI assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "httpx>=0.28,<0.29",
    "uvicorn[standard]>=0.36,<0.37",
    "openai>=1.40",
    "openai-chatkit>=1.0.2,<2",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
